# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all developers
      parameters:
        - name: skill
          in: query
          schema:
            type: string
          description: Filter by skill/tech stack
        - name: search
          in: query
          schema:
            type: string
          description: Search by username or display name
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      summary: Get user profile by username
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/projects:
    get:
      operationId: getUserProjects
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      summary: Get projects for a user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      summary: Get blog posts for a user
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: tech
          in: query
          schema:
            type: string
          description: Filter by tech stack tag
        - name: search
          in: query
          schema:
            type: string
          description: Search by title or description
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get a project by ID
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Delete a project
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        githubUsername:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, username, createdAt, projectCount, postCount]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        githubUsername:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        projectCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        githubUsername:
          type: string
        website:
          type: string
        skills:
          type: array
          items:
            type: string
    Project:
      type: object
      required: [id, userId, title, createdAt, likeCount, saveCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        likeCount:
          type: integer
        saveCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      required: [projects, total, page, limit]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProjectInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        imageUrl:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        imageUrl:
          type: string
    Post:
      type: object
      required: [id, userId, title, published, createdAt, likeCount, saveCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        category:
          type: ["string", "null"]
        published:
          type: boolean
        likeCount:
          type: integer
        saveCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...